move movepositions to positions
