+++ src/main/java/frc/robot/commands/sequences/BaseAutonSequence.java
@@ -11,6 +11,8 @@ import frc.robot.commands.grabber.AidenIntakeCommand;
 import frc.robot.commands.grabber.AidenPlaceCommand;
 import frc.robot.commands.mover.JulianLevelCommand;
 import frc.robot.commands.swerve.FollowPathCommand;
+import frc.robot.positions.PiecePosition;
+import frc.robot.positions.PlacePosition;
 import frc.robot.subsystems.RollerSubsystem;
 import frc.robot.subsystems.TiltedElevatorSubsystem;
 import frc.robot.subsystems.TiltedElevatorSubsystem.ElevatorState;
