+++ src/main/java/frc/robot/commands/sequences/BaseBalanceAutonSequence.java
@@ -6,6 +6,7 @@ import edu.wpi.first.math.geometry.Pose2d;
 
 import frc.robot.commands.BalancerCommand;
 import frc.robot.commands.swerve.FollowPathCommand;
+import frc.robot.positions.PlacePosition;
 import frc.robot.subsystems.RollerSubsystem;
 import frc.robot.subsystems.TiltedElevatorSubsystem;
 import frc.robot.subsystems.drivetrain.BaseSwerveSubsystem;
