+++ src/main/java/frc/robot/commands/sequences/BaseBottomAutonSequence.java
@@ -6,6 +6,8 @@ import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.wpilibj2.command.Command;
 
 import frc.robot.commands.swerve.FollowPathCommand;
+import frc.robot.positions.PiecePosition;
+import frc.robot.positions.PlacePosition;
 import frc.robot.subsystems.RollerSubsystem;
 import frc.robot.subsystems.TiltedElevatorSubsystem;
 import frc.robot.subsystems.drivetrain.BaseSwerveSubsystem;
