+++ src/main/java/frc/robot/commands/sequences/BlueTopAuton.java
@@ -3,7 +3,8 @@ package frc.robot.commands.sequences;
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
 import edu.wpi.first.math.util.Units;
-
+import frc.robot.positions.PiecePosition;
+import frc.robot.positions.PlacePosition;
 import frc.robot.subsystems.RollerSubsystem;
 import frc.robot.subsystems.TiltedElevatorSubsystem;
 import frc.robot.subsystems.drivetrain.BaseSwerveSubsystem;
