+++ src/main/java/frc/robot/positions/PiecePosition.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.sequences;
+package frc.robot.positions;
 
 import edu.wpi.first.math.geometry.Pose2d;
 import edu.wpi.first.math.geometry.Rotation2d;
